Add cast in isdigit() argument. Necessary e.g. for sparc-solaris.
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1014 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/parser/text.c
@@ -1606,7 +1606,7 @@ int GWEN_Text_UnescapeXmlToBuffer(const char *src, GWEN_BUFFER *buf) {
 	unsigned char num=0;
 
 	src++;
-	while(isdigit(*src)) {
+	while(isdigit((int)*src)) {
 	  num*=10;
 	  num+=(*src)-'0';
 	  src++;
